Add makefile (not tested on Pi)
to use for nonpi compilation use `make test`
