+++ src/driver/wiringPi_test_env.h
@@ -11,8 +11,7 @@
 
 #ifndef WIRINGPI_TEST_ENV_H
 #define WIRINGPI_TEST_ENV_H
-#include <iostream>
-#include <stdlib.h>
+
 
 #define HIGH   1
 #define LOW    0
@@ -20,25 +19,9 @@
 #define INPUT  0
 
 
-int wiringPiSetupGpio()
-{
-    return 0;
-}
-void pinMode( int p1, int p2 )
-{
-    std::cout << "Set pinMode: pin " << p1 << " mode: " << p2 << std::endl;
-}
-
-int digitalRead( int p )
-{
-    int read_val = rand() > RAND_MAX / 2;
-    std::cout << "Read pin:  " << p << " value: " << read_val << std::endl;
-    return read_val;
-}
-
-void digitalWrite( int pin, int val )
-{
-    std::cout << "Write pin: " << pin << " value: " << val << std::endl;
-}
+int  wiringPiSetupGpio();
+void pinMode( int p1, int p2 );
+int  digitalRead( int p );
+void digitalWrite( int pin, int val );
 
 #endif  // WIRINGPI_TEST_ENV_H
